fix wrong size argument on string copy
The size argument to strlcpy() needed to be subtracted by one since the
starting array is offset by one. The unchecked return value also leaves
a question to the readers on weather truncation is desired/possible or
not.

Instead use memccpy() with the correct size, and assert() that
truncation cannot occur since the basename cannot be larger than the
absolute path.

This additionally avoids dependency on libbsd.